third_party/blink/web_tests/external/wpt/css/css-properties-values-api/at-property-shadow-expected.txt

This is a testharness.js-based test.
FAIL @property rules in shadow trees should have no effect assert_equals: expected "calc(1px + 1px)" but got " calc(1px + 1px)"
Harness: the test ran to completion.
